fix rf and svm on-device inference: quantization/decision-function bugs
Both classic models were essentially broken on the ESP32 despite good
notebook results, for two unrelated reasons:

- RF: emlearn.convert(..., method="inline") defaults to dtype='int16_t',
  which truncates each tree split threshold to int(value) with no
  scaling. ECG feature magnitudes here are almost all well under 1.0, so
  nearly every threshold collapsed to -1/0/1/2 -- verified in the
  checked-in header (random_forest_classifier.h had thresholds like
  "features[2] < -1"). This is consistent with the on-device benchmark's
  exact 0.2/0.8 macro sensitivity/specificity, the signature of a
  classifier that always outputs the same class. Re-exported with
  dtype="float" (verified against the pickled sklearn model: the old
  int16 header disagreed with sklearn's own predict() a lot more than
  the new float header does on synthetic inputs) and dropped the
  firmware's *1000 int16 quantization step, which only made sense for
  the old scheme.

- SVM: the generated svm_predict() summed every support vector against a
  single dual_coef row per one-vs-one pair, instead of libsvm's actual
  formula -- each pair (i, j) only uses dual_coef_[j-1] over class i's
  support vectors and dual_coef_[i] over class j's, with the kernel
  value against each support vector computed once and reused across
  pairs. Rewrote export_classifier_header() to match libsvm's
  svm_predict_values() exactly (verified bit-for-bit against
  svc.predict() on 600 synthetic/support-vector inputs: 100% match,
  versus the old formula which was simply wrong). This also cuts
  inference from ~500ms to roughly 1/10th, since each pair now only
  touches its own support vectors instead of all ~13k. kvalue is a
  static array, not a stack one, since 13k+ floats would blow the ESP32
  task stack.

Regenerated random_forest_classifier.h and svm_classifier.h from the
already-fitted models in ml/models/*.pickle|joblib -- no retraining
needed, since these were export-time bugs, not training-time ones.

+++ ml/src/classifiers/random_forest_classifier.py
@@ -20,7 +20,11 @@ def export_model(forest_classifier):
     output_directory = Path('models')
     output_directory.mkdir(parents=True, exist_ok=True)
 
-    c_model = emlearn.convert(forest_classifier, method="inline")
+    # dtype='float' avoids emlearn's default int16 quantization, which
+    # truncates each split threshold to int(value) with no scaling -- for
+    # small-magnitude ECG statistics (means, stds well under 1.0) that
+    # collapses almost every threshold to -1/0/1/2, destroying the tree.
+    c_model = emlearn.convert(forest_classifier, method="inline", dtype="float")
     c_model.save(file=str(output_directory / "random_forest_classifier.h"), name="random_forest")
     print('[EXPORT] Saved Random Forest C header model.')
